+++ app/src/main/java/com/bn/todo/ui/MainActivity.kt
@@ -14,8 +14,8 @@ import com.bn.todo.data.model.TodoList
 import com.bn.todo.databinding.ActivityMainBinding
 import com.bn.todo.ktx.*
 import com.bn.todo.ui.callback.TodoClickCallback
-import com.bn.todo.ui.view.TodoInfoFragment
-import com.bn.todo.ui.viewmodel.TodoListViewModel
+import com.bn.todo.ui.main.view.TodoInfoFragment
+import com.bn.todo.ui.main.viewmodel.TodoListViewModel
 import com.bn.todo.util.DialogUtil
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.flow.combine
