+++ app/src/main/res/layout/fragment_create_todo.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".ui.view.CreateTodoFragment">
+    tools:context=".ui.main.view.CreateTodoFragment">
 
     <androidx.constraintlayout.widget.Guideline
         android:id="@+id/guideline"
