+++ app/src/main/res/layout/fragment_todo_info.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:padding="8dp"
-    tools:context=".ui.view.TodoInfoFragment">
+    tools:context=".ui.main.view.TodoInfoFragment">
 
     <TextView
         android:id="@+id/title_text"
