+++ app/src/test/java/com/bn/todo/ui/main/viewmodel/TodoViewModelTest.kt
@@ -1,4 +1,4 @@
-package com.bn.todo.ui.viewmodel
+package com.bn.todo.ui.main.viewmodel
 
 import app.cash.turbine.test
 import com.bn.todo.MainCoroutineExtension
